User registration and validation

+++ base/views.py
@@ -5,6 +5,7 @@ from django.contrib import messages
 from django.contrib.auth import authenticate, login, logout
 from django.contrib.auth.decorators import login_required
 from django.http import HttpResponse
+from django.contrib.auth.forms import UserCreationForm
 from .models import Room, Topic
 from .forms import RoomForm
 
@@ -18,6 +19,12 @@ from .forms import RoomForm
 # ]
 
 def loginPage(request):
+
+    page = 'login'
+
+    #Saving user details after they are already logged in
+    if request.user.is_authenticated:
+        return redirect('home')
 
     if request.method == 'POST':
         username = request.POST.get('username')
@@ -37,13 +44,24 @@ def loginPage(request):
         else:
             messages.error(request, 'Username OR Password does not exist')
 
-    context = {}
+    context = {'page': page}
     return render(request, 'base/login_register.html', context)
 
 
 def logoutUser(request):
     logout(request)
     return redirect('home')
+
+def registerPage(request):
+    form = UserCreationForm() 
+
+    if request.method == 'POST':
+        form = UserCreationForm(request.POST)
+        if form.is_vaild():
+            user = form.save(commit=False)
+            user.username = user.username.lower()
+            
+    return render(request, 'base/login_register.html', {'form': form})
 
 def home(request):
     q = request.GET.get('q') if request.GET.get('q') != None else '' #q is going to be equal to whatever we passed in the url
@@ -93,7 +111,7 @@ def updateRoom(request, pk):
     #When the values don't match then this is not going to work
     
     if request.user != room.host:
-        return HttpResponse('You are not Logged in!') 
+        return HttpResponse('You are not allowed here!') 
 
     if request.method == 'POST':
         form = RoomForm(request.POST, instance=room) #This is going to tell our function what room to update
@@ -107,6 +125,10 @@ def updateRoom(request, pk):
 @login_required(login_url='login')
 def deleteRoom(request, pk):
     room = Room.objects.get(id=pk)
+
+    if request.user != room.host:
+        return HttpResponse('You are not allowed here!') 
+
     if request.method == 'POST':
         room.delete()
         return redirect('home')
